2005-01-31  Kazuki Oikawa <kazuki@panicode.com>
	* Patch from Kazuki Oikawa <kazuki@panicode.com> that replaces the
	old Integer and Floating point number string formatter with a new
	formatter that is more compliant. 

This fixes: http://bugzilla.ximian.com/show_bug.cgi?id=71286


svn path=/trunk/mcs/; revision=39817

+++ mcs/class/corlib/System/SingleFormatter.cs
@@ -45,7 +45,7 @@ namespace System {
 		const int dec_len2 = 7;
 		const int dec_len_min2 = -16;
 
-		public static string NumberToString (string format, NumberFormatInfo nfi, double value)
+		public static string NumberToString (string format, NumberFormatInfo nfi, float value)
 		{
 			if (fpf == null) {
 				fpf = new FloatingPointFormatter (p, p10, dec_len, dec_len_min, p2, p102, dec_len2, dec_len_min2);
